Verified read functions work properly
